Move Manifest attributes to specific files
